removed my social links from pmpermit

+++ stdplugins/pmpermit.py
@@ -6,7 +6,7 @@ import json
 from telethon import events
 from telethon.tl import functions, types
 from sql_helpers.pmpermit_sql import is_approved, approve, disapprove, get_all_approved
-from sql_helpers.global_variables_sql import SYNTAX, MODULE_LIST
+from sql_helpers.global_variables_sql import SYNTAX, MODULE_LIST,SUDO_USERS
 from uniborg.util import admin_cmd
 
 
@@ -19,8 +19,9 @@ UNIBORG_USER_BOT_WARN_ZERO = "```Blocked! Thanks for the spam.```"
 UNIBORG_USER_BOT_NO_WARN = """```
 Hee HAA! This is a bot. Don't fret.\nMy master hasn't approved you to PM.\nPlease
  wait for my master to look in, he mostly approves PMs.\n
-MeanWhile You can Check My Creator(s)  Social Accounts. :)```\n
-[@Instagram](https://www.instagram.com/your_._stranger)\n\n[@Github](https://github.com/authoritydmc)\n\n[@Facebook](https://www.facebook.com/AuthorityDMC)
+
+```\n
+
 """
 
 @borg.on(events.NewMessage(incoming=True, func=lambda e: e.is_private))
@@ -33,11 +34,16 @@ async def monito_p_m_s(event):
         if not is_approved(chat.id) and chat.id != borg.uid:
             logger.info(chat.stringify())
             logger.info(borg.storage.PM_WARNS)
+            if chat.id in SUDO_USERS:
+                print("Welcome Master.")
+                approve(chat.id,"SUDO USER")
+                return
             if chat.id not in borg.storage.PM_WARNS:
                 borg.storage.PM_WARNS.update({chat.id: 0})
             if borg.storage.PM_WARNS[chat.id] == Config.MAX_FLOOD_IN_P_M_s:
                 r = await event.reply(UNIBORG_USER_BOT_WARN_ZERO)
                 await asyncio.sleep(3)
+                
                 await borg(functions.contacts.BlockRequest(chat.id))
                 if chat.id in borg.storage.PREV_REPLY_MESSAGE:
                     await borg.storage.PREV_REPLY_MESSAGE[chat.id].delete()
